Add header to find_pcr_product, separate header tests
