+++ tests/test_Product_functions.py
@@ -124,60 +124,3 @@ class TestFindProduct:
         )
 
         assert expected_results == actual_results
-
-
-class TestHeaders:
-    @pytest.fixture(scope="class")
-    def primers(self) -> Iterator[List[FastaSequence]]:
-        forward_primer = FastaSequence("test_forward", "GGAG")
-        reverse_primer = FastaSequence("test_reverse", "TAAT")
-        yield [forward_primer, reverse_primer]
-
-    @pytest.fixture(scope="class")
-    def small_sequence_1(self) -> Iterator[FastaSequence]:
-        sequence = FastaSequence("test_sequence", "GGAGCATGCTATGTCGTAGCTGATGCAATTA")
-        yield sequence
-
-    @pytest.fixture(scope="class")
-    def medium_sequence_1(self) -> Iterator[FastaSequence]:
-        seq_string = "AAAACCCACCCAAAAAATATTCTTTTGCATCCACTGTCAACTTTTCACAGAAACCCATTAAGTCAGGATCCTTAAGAGTTTCCGAGTGTTCATCTGCTGATATTCCAACAACAAACTCTACCGAGTGTCTGAATTTGCTGCTTGAAAAGAGAGGAGCTTCATCGAGTCAAAACTGTTGGAGAAAGATTTCTCTTGAAGATCTTTTCTGTTCCACTTCAAACCTTCCTTCCCCTACTAAAGGGAATCTCCCAATTATTGCCGACGCTGGTGACCATGGGATTTTATCTTTCAAAGTTTCTGACCTGAAAGAAGATATACCCTCGCAAATATCGACGGCTAAGGAAGAGTCCTTCAGTGGTAATGAAGAAGAAGAAGAAGAAGAAGGTGATGACGATGATAAGATAACCCTTCAGGATTTTGTTTGTAATGAAAAGAACCAAAAAGAAATGGGTGAACAAAGAAATGACGTAAGCTCGTCTTCTTGGGTACAAACTGAGCTGTTGTTTCTTCTCCTAAAGGGAAGTATCGGGAGTAATGATACTCAAACAACACTGAGAAAACCAACCCTGTTTCTGATTCCACATTA"
-        sequence = FastaSequence("medium_test_sequence", seq_string)
-        yield sequence
-
-    def test_no_header(
-        self, primers: List[FastaSequence], small_sequence_1: FastaSequence
-    ) -> None:
-        expected_results = ""
-
-        forward_primer, reverse_primer = primers
-        pcr_results = get_pcr_product(
-            sequence=small_sequence_1,
-            forward_primer=forward_primer,
-            reverse_primer=reverse_primer,
-            header=False,
-        ).splitlines()
-        actual_results = "".join(
-            [line for line in pcr_results if "forward_primer" in line]
-        )
-
-        assert expected_results == actual_results
-
-    def test_basic_header(
-        self, primers: List[FastaSequence], small_sequence_1: FastaSequence
-    ) -> None:
-        expected_results = (
-            "forward_primer\treverse_primer\tstart\tend\tlength\tsequence"
-        )
-
-        forward_primer, reverse_primer = primers
-        pcr_results = get_pcr_product(
-            sequence=small_sequence_1,
-            forward_primer=forward_primer,
-            reverse_primer=reverse_primer,
-            header=True,
-        ).splitlines()
-        actual_results = "".join(
-            [line for line in pcr_results if line == expected_results]
-        )
-
-        assert expected_results == actual_results
